requirements.txt
aiomysql==0.0.22
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
black==22.1.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
databases==0.5.5
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
docopt==0.6.2
ecdsa==0.17.0
entrypoints==0.4
executing==0.8.2
fastapi==0.73.0
fastapi-login==1.7.3
fastapi-sessions==0.3.2
Flask==2.0.2
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.26
googleapis-common-protos==1.54.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonpickle==1.5.2
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lab==7.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
munch==2.5.0
mypy-extensions==0.4.3
mysql-connector==2.2.9
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathlib2==2.3.7.post1
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.27
protobuf==3.17.3
pure-eval==0.2.2
py-cpuinfo==8.0.0
py4j==0.10.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyJWT==2.3.0
PyMySQL==0.9.3
pyodbc==4.0.32
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytimeparse==1.1.8
pytz==2021.3
pywin32==303
pywinpty==1.1.6
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
redis==4.1.3
requests==2.27.1
rsa==4.8
sacred==0.8.2
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.4.31
stack-data==0.1.4
starlette==0.17.1
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
txt2tags==3.7
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.8
uvicorn==0.17.4
verta==0.19.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wrapt==1.13.3
